Hibana.Controller (Hibana v0.1.0)

Base controller module for handling HTTP responses.

Usage

defmodule MyApp.UserController do
  use Hibana.Controller

  def index(conn) do
    json(conn, %{users: []})
  end

  def show(conn) do
    id = conn.params["id"]
    json(conn, %{user: %{id: id, name: "User"}})
  end
end

Response Helpers

json/1

Sends a JSON response with application/json content type.

json(conn, %{message: "Hello!"})

html/1

Sends an HTML response with text/html content type.

html(conn, "<h1>Hello!</h1>")

text/1

Sends a plain text response with text/plain content type.

text(conn, "Hello, World!")

redirect/1

Redirects to another URL.

redirect(conn, to: "/users")

send_file/2

Sends a file as a download.

send_file(conn, "/path/to/file.pdf")

Parameter Access

conn.params - Path parameters (e.g., /users/:id)

conn.body_params - Request body as a map

conn.query_params - Query string parameters

conn.req_headers - Request headers

Session Helpers

get_session(conn, key) - Get session value

put_session(conn, key, value) - Set session value

Functions

fetch_body_params(conn, parsers \\ [])

Fetch and parse body parameters from the connection.

fetch_query_params(conn, opts \\ [])

Fetch and parse query parameters from the connection.

get_body_params(conn)

Get the parsed body parameters from the connection.

get_query_params(conn)

Get the query string parameters from the connection.

get_session(conn, key)

Get a session value by key.

get_status(conn)

Get the current HTTP status code from the connection.

html(conn, content)

Send an HTML response.

json(conn, data)

Send a JSON response with the given data.

put_session(conn, key, value)

Store a value in the session under the given key.

put_status(conn, status)

Set the HTTP status code on the connection.

redirect(conn, list)

Redirect to the given URL with a 302 status.

render(conn, view, template, assigns \\ %{})

Render a template using the given view module and send as HTML.

req_header(conn, key)

Get the first value of a request header by name (case-insensitive).

send_file(conn, path, opts \\ [])

Send a file as a download response with optional filename and content type.

text(conn, content)

Send a plain text response.