Hibana.Features (Hibana v0.1.0)

Feature toggle system for enabling/disabling plugins and modules via config.

Usage

In config/config.exs, list which features are enabled:

config :hibana, :features, [
  Hibana.Plugins.Logger,
  Hibana.Plugins.BodyParser,
  Hibana.Plugins.CORS,
  Hibana.Plugins.JWT,
  # Hibana.Plugins.RateLimiter,    # <-- commented = disabled
  # Hibana.Plugins.GraphQL,        # <-- commented = disabled
]

Or use a map for explicit on/off:

config :hibana, :features, %{
  Hibana.Plugins.Logger => true,
  Hibana.Plugins.BodyParser => true,
  Hibana.Plugins.CORS => true,
  Hibana.Plugins.RateLimiter => false,   # disabled
  Hibana.Plugins.GraphQL => false,        # disabled
}

In your Router

defmodule MyApp.Router do
  use Hibana.CompiledRouter
  import Hibana.Features

  # Only loads if enabled in config
  feature_plug Hibana.Plugins.Logger
  feature_plug Hibana.Plugins.BodyParser
  feature_plug Hibana.Plugins.CORS, origins: ["*"]
  feature_plug Hibana.Plugins.RateLimiter, max_requests: 100
  feature_plug Hibana.Plugins.JWT, secret: "secret"

  get "/", PageController, :index
end

Runtime Check

if Hibana.Features.enabled?(Hibana.Plugins.RateLimiter) do
  # rate limiter logic
end

Supervisor Children

# Only start enabled services
children = Hibana.Features.filter_children([
  {Hibana.Cluster, strategy: :epmd},
  {Hibana.PersistentQueue, name: :jobs},
  {Hibana.Plugins.Metrics, []},
])

Functions

disable(module)

Disable a feature at runtime

enable(module)

Enable a feature at runtime

enabled?(module)

Check if a feature/module is enabled in config

feature_plug(module, opts \\ [])
(macro)

Conditionally add a plug only if the module is enabled in config

filter_children(children)

Filter a list of supervisor children, keeping only enabled ones

list_disabled()

List all disabled features

list_enabled()

List all enabled features